requirements.txt
# IBKR connection
ib_insync==0.9.86

# Technical analysis
ta==0.11.0
pandas==2.2.3
numpy==1.26.4

# Dashboard
streamlit==1.41.1
plotly==5.24.1

# Notifications
python-telegram-bot==21.9

# Backtesting data
yfinance==0.2.51
pyarrow==18.1.0

# Utilities
python-dotenv==1.0.1
pytz==2024.2
aiohttp==3.11.11
aiosqlite==0.20.0
